Update Classes and update structuring
